* commit '013c9c324e281857ca584b7d092026b3062aedca':
  avcenc: Switch malloc/free callbacks to use pointers instead of ints

+++ media/libstagefright/codecs/avc/enc/src/motion_est.cpp
@@ -176,7 +176,7 @@ void CleanMotionSearchModule(AVCHandle *avcHandle)
 
     if (encvid->mvbits_array)
     {
-        avcHandle->CBAVC_Free(avcHandle->userData, (int)(encvid->mvbits_array));
+        avcHandle->CBAVC_Free(avcHandle->userData, encvid->mvbits_array);
         encvid->mvbits = NULL;
     }
